+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.22.0] - 2019-05-29
+
+### Added
+- Added the `--list` option to list all the tasks in the toastfile.
+- Added the `description` task field to be shown to the user when `--list` is used.
+
 ## [0.21.0] - 2019-05-26
 
 ### Added